- Changes to NR spike sort (new projection method) - Dealing with a very strange CL issue: longer loop times actually are more problematic than short ones for the input side of things. It is more likely that with a longer loop time, each time NR tries to inject data into datsrv, it will take a long time (blocks read threads) and thus causes a larger lag on the output side of things. I have not fully worked out what is going on there.
git-svn-id: https://neurorighter.googlecode.com/svn/trunk@409 471701b0-e017-11dd-9759-139c9ab14fc6

+++ simoc/spk2obs/Spk2ASDR.cs
@@ -24,8 +24,9 @@ namespace simoc.spk2obs
 
         internal override void MeasureObservable()
         {
-            // Estimate the CSDR
-            currentObservation = (double)newSpikes.eventBuffer.Count() / (double)numSecondInCurrentRead; 
+            // Estimate the ASDR
+            currentObservation = (double)newSpikes.eventBuffer.Count() / numSecondInCurrentRead;
+            //currentObservation = numSecondInCurrentRead; 
         }
